fix: add theme data source for devtools.
